A    hermit-fix/src/edu/ucla/sspace/tools/RandomSemanticSpaceGenerator.java   - A new semantic space generator that randomly swaps and permutes semantic     vectors from a provided semantic space.
M    hermit-fix/src/edu/ucla/sspace/vector/Vectors.java
  - Add a special case for CompactSparseVectors

M    hermit-fix/src/edu/ucla/sspace/index/DefaultPermutationFunction.java
  - Avoid an instance of auto-boxing

M    hermit-fix/src/edu/ucla/sspace/index/DoubleVectorGenerator.java
  - Made generic

A    hermit-fix/src/edu/ucla/sspace/index/RandomOrthogonalVectorGenerator.java
  - A new vector generator that creates random orthogonal vectors using the
    gramm-schmidt process

+++ branches/hermit-fix/src/edu/ucla/sspace/index/DefaultPermutationFunction.java
@@ -171,7 +171,7 @@ public class DefaultPermutationFunction
         }
 
         for (int d : dimensions)
-            result.set(d, v.getValue(d).intValue());
+            result.set(d, v.getValue(d));
 
         return result;
     }
